backend/requirements.txt
Django>=4.2,<6.1
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-filter>=23.5,<26.0
django-cors-headers>=4.3,<5.0
drf-spectacular>=0.27,<0.30
channels>=4.0,<5.0
channels-redis>=4.2,<5.0
Pillow>=10.0,<12.0
psycopg>=3.1,<4.0